requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.7.0
idna==3.10
inflection==0.5.1
pillow==11.0.0
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uuid==1.30
